Move the source code directory back to “floris”
