+++ floris/types.py
