+++ tests/turbine_map_test.py
@@ -15,9 +15,9 @@ import numpy as np
 import copy
 import pytest
 from .sample_inputs import SampleInputs
-from src.floris.types import Vec3
-from src.floris.turbine import Turbine
-from src.floris.turbine_map import TurbineMap
+from floris.types import Vec3
+from floris.turbine import Turbine
+from floris.turbine_map import TurbineMap
 
 class TurbineMapTest():
     def __init__(self):
